LPD-58417 Merge DeleteIndexRequestExecutor/DeleteIndexRequestExecutorImpl into pojos
