+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/search/engine/adapter/index/ElasticsearchIndexRequestExecutor.java
@@ -155,6 +155,8 @@ public class ElasticsearchIndexRequestExecutor implements IndexRequestExecutor {
 	protected void activate() {
 		_analyzeIndexRequestExecutor = new AnalyzeIndexRequestExecutor(
 			_elasticsearchClientResolver);
+		_deleteIndexRequestExecutor = new DeleteIndexRequestExecutor(
+			_elasticsearchClientResolver);
 		_getMappingIndexRequestExecutor = new GetMappingIndexRequestExecutor(
 			_elasticsearchClientResolver);
 		_openIndexRequestExecutor = new OpenIndexRequestExecutor(
@@ -171,7 +173,6 @@ public class ElasticsearchIndexRequestExecutor implements IndexRequestExecutor {
 	@Reference
 	private CreateIndexRequestExecutor _createIndexRequestExecutor;
 
-	@Reference
 	private DeleteIndexRequestExecutor _deleteIndexRequestExecutor;
 
 	@Reference
