--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/DeleteIndexRequestExecutorImpl.java
@@ -1,25 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.search.engine.adapter.index;
-
-import com.liferay.portal.search.engine.adapter.index.DeleteIndexRequest;
-import com.liferay.portal.search.engine.adapter.index.DeleteIndexResponse;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Bryan Engler
- */
-@Component(service = DeleteIndexRequestExecutor.class)
-public class DeleteIndexRequestExecutorImpl
-	implements DeleteIndexRequestExecutor {
-
-	@Override
-	public DeleteIndexResponse execute(DeleteIndexRequest deleteIndexRequest) {
-		throw new UnsupportedOperationException();
-	}
-
-}
