+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/SolrIndexRequestExecutor.java
@@ -156,8 +156,8 @@ public class SolrIndexRequestExecutor implements IndexRequestExecutor {
 	@Reference
 	private CreateIndexRequestExecutor _createIndexRequestExecutor;
 
-	@Reference
-	private DeleteIndexRequestExecutor _deleteIndexRequestExecutor;
+	private final DeleteIndexRequestExecutor _deleteIndexRequestExecutor =
+		new DeleteIndexRequestExecutor();
 
 	@Reference
 	private FlushIndexRequestExecutor _flushIndexRequestExecutor;
